chore: 清理冗余文件，整合项目结构
- 删除 SubspaceParasite/main/ (旧版代码副本，101 Java文件)
- 删除 srp_framework/ (最旧版本，文档已保存至docs/)
- 删除 src_backup/ (旧备份)
- 删除 srp_original/ (空目录)
- 删除 MCMOS_extracted/ (已集成到resources，节省48M)
- 将根目录 src/ 合并到 SubspaceParasite/src/ (更完整的代码)
- 整理文档到 docs/planning/ 和 docs/framework-reports/
- 整理工具脚本到 tools/
- 项目体积从312M减少到169M

+++ SubspaceParasite/src/main/java/com/subspaceparasite/common/entity/base/EvolutionDispatcher.java
@@ -278,9 +278,9 @@ public class EvolutionDispatcher {
         int evoLevel = parasite.getEvolutionComponent().getEvolutionLevel();
         int evoPoints = (int) parasite.getEvolutionComponent().getEvolutionPointsInternal();
         
-        // Send sync packet
+        // Send sync packet via ModNetwork
         ModNetwork.CHANNEL.sendToAllPlayers(
-            new ModNetwork.EvolutionSyncPacket(entityId, evoLevel, evoPoints),
+            new ModNetwork.EvolutionSyncPacket(entityId, evoLevel, (float)evoPoints),
             (ServerLevel) parasite.level()
         );
     }
